Initial commit
These are the files used in the cron job that creates thumbnails of
images collected by WiscScan and copies them to the Forward directory on
open.microscopy.wisc.edu. In order to run, CreateThumbnail.java must be
compiled with loci_tools.jar in the classpath and the log4j.properties
file. Paths in the "publish" script are relative to the
open.microscopy.wisc.edu machine.